Latest progress on importing and printing words.
Initial check of anagram_solver.h and program.c
Removed main() from anagram_solver.c and placed it in program.c.

Re-organized comments and docstrings for anagram_solver source and header
files. The is_supported() function now has a newline count check to
eliminate unsupported words.

Initial draft the build() function for anagram_solver.c is complete.
The solve() function is currently functional and outputs words after
searching in a hash table, but the words are incorrect. Additional
checking needs to be put in place to verify whether two words are truly
anagrams or not.