Anagram Solver backend is now functional.
Anagram Solver's solve() function now solves anagrams. Some UI needs to be
built on top of it. The solve() function is also now case-agnostic.

The length of a word has been added as a member in linked list nodes. This
is used to ignore words when solving for anagrams.

Modified files need clean up.